+++ VKApp/VKApp/Models/Photo/ItemPhoto.swift
@@ -1,21 +1,8 @@
-// Photo.swift
+// ItemPhoto.swift
 // Copyright © RoadMap. All rights reserved.
 
 import Foundation
 import RealmSwift
-
-/// Фотография
-final class Photo: Decodable {
-    let response: ResponsePhoto
-}
-
-/// Ответ с сервера о фотографии
-final class ResponsePhoto: Decodable {
-    /// Количество фотографий
-    let count: Int
-    /// Фотографии
-    let items: [ItemPhoto]
-}
 
 /// Фотография пользователя
 final class ItemPhoto: Object, Decodable {
