+++ VKApp/VKApp/Models/Photo/Photo.swift
@@ -1,0 +1,10 @@
+// Photo.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+import RealmSwift
+
+/// Фотография
+final class Photo: Decodable {
+    let response: ResponsePhoto
+}
